Fixing warnings

+++ src/bot/bot_api.cpp
@@ -26,8 +26,8 @@ void fill_entity_struct(const ::model::components &components, ninja_api::nnj_en
 			entity->properties.rotation_speed = properties.rotation_speed;
 			entity->properties.attack_range   = properties.attack_range;
 			entity->properties.activate_range = properties.activate_range;
-			entity->properties.attack_delay   = properties.attack_delay;
-			entity->properties.throw_delay    = properties.throw_delay;
+			entity->properties.attack_delay   = static_cast<float>(properties.attack_delay);
+			entity->properties.throw_delay    = static_cast<float>(properties.throw_delay);
 
 			if (components.state[entity->handle].preparing_action.has_value()) {
 				entity->state = ninja_api::nnj_entity_state::ES_BUSY;
